service_type 7 output & ignore monitor_state 0
The output for service_type 7 didn't work in all cicumstances: `element.findall('program/output')` sometimes is an empty list and its first non-existent entry is of the type None. If the list is not empty get the last xml.etree.ElementTree.Element as text from it as the return value. If it is empty return the message 'no command output available'.

Instead of only taking in account monitor_states 1 (Yes) and 2 (Init) take all monitor_states in account that are not 0 (Not). There is no reason not to include information about any monitor that is not disabled.